experiments/requirements.txt
pysen
isort
mypy
black
flake8
aiofiles
aiohttp
beautifulsoup4
loguru
lxml
openai
requests
streamlit
types-aiofiles
types-requests
types-PyYAML
faiss-cpu
python-dotenv
torch
transformers
weave
